+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version: 6.0
 import PackageDescription
 
 
@@ -16,8 +16,8 @@ let package = Package(
     targets: [
       .binaryTarget(
         name: "Tpay",
-        url: "https://github.com/tpay-com/tpay-ios/releases/download/1.3.3/Tpay.xcframework.zip",
-        checksum: "5bcfdb2fa75205dafff677e97d7a4bd97a71f99d80e2823fc1a2035db99d246d"
+        url: "TO_BE_REPLACED_DURING_CI_BUILD",
+        checksum: "TO_BE_REPLACED_DURING_CI_BUILD"
       )
     ]
 )