+++ Tpay/Tpay/Dependencies/Kingfisher/AnimatedImageView.swift
@@ -427,7 +427,7 @@ extension AnimatedImageView {
         }
 
         /// The index of the current animation frame.
-        internal(set) var currentFrameIndex = 0 {
+        var currentFrameIndex = 0 {
             didSet {
                 previousFrameIndex = oldValue
             }