+++ Tpay/Tpay/Dependencies/Kingfisher/KingfisherManager.swift
@@ -757,7 +757,7 @@ class CacheCallbackCoordinator {
     private let stateQueue: DispatchQueue
     private var threadSafeState: State = .idle
 
-    private (set) var state: State {
+    private(set) var state: State {
         set { stateQueue.sync { threadSafeState = newValue } }
         get { stateQueue.sync { threadSafeState } }
     }