+++ Tpay/Tpay/Dependencies/Snail/UIViewControllerExtensions.swift
@@ -9,11 +9,11 @@ extension UIViewController {
     private static var disposerKey = "com.compass.Snail.UIViewController.disposer"
 
     var disposer: Disposer {
-        if let disposer = objc_getAssociatedObject(self, &UIViewController.disposerKey) as? Disposer {
+        if let disposer = objc_getAssociatedObject(self, UIViewController.disposerKey.pointer) as? Disposer {
             return disposer
         }
         let disposer = Disposer()
-        objc_setAssociatedObject(self, &UIViewController.disposerKey, disposer, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
+        objc_setAssociatedObject(self, UIViewController.disposerKey.pointer, disposer, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
         return disposer
     }
 }