+++ Tpay/Tpay/Sources/Internal/Extensions/UIScrollView+Extensions.swift
@@ -10,7 +10,7 @@ extension UIScrollView {
     private static var contentOffsetObservableKey = "com.tpay.sdk.UIScrollView.ContentOffsetObservableKey"
     
     private var contentOffsetObservationToken: NSKeyValueObservation {
-        if let observationToken = objc_getAssociatedObject(self, &UIScrollView.contentOffsetObservationTokenKey) as? NSKeyValueObservation {
+        if let observationToken = objc_getAssociatedObject(self, UIScrollView.contentOffsetObservationTokenKey) as? NSKeyValueObservation {
             return observationToken
         }
         let observationToken = observe(\.contentOffset, options: .new) { [weak self] _, change in
@@ -19,17 +19,17 @@ extension UIScrollView {
             }
             self?.contentOffsetObservable.on(.next(newValue))
         }
-        objc_setAssociatedObject(self, &UIScrollView.contentOffsetObservationTokenKey, observationToken, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
+        objc_setAssociatedObject(self, UIScrollView.contentOffsetObservationTokenKey, observationToken, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
 
         return observationToken
     }
     
     private var _contentOffsetObservable: Observable<CGPoint> {
-        if let observable = objc_getAssociatedObject(self, &UIScrollView.contentOffsetObservableKey) as? Observable<CGPoint> {
+        if let observable = objc_getAssociatedObject(self, UIScrollView.contentOffsetObservableKey) as? Observable<CGPoint> {
             return observable
         }
         let observable = Observable<CGPoint>()
-        objc_setAssociatedObject(self, &UIScrollView.contentOffsetObservableKey, observable, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
+        objc_setAssociatedObject(self, UIScrollView.contentOffsetObservableKey, observable, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
         return observable
     }
     
@@ -45,7 +45,7 @@ extension UIScrollView {
     private static var contentSizeObservableKey = "com.tpay.sdk.UIScrollView.ContentSizeObservableKey"
     
     private var contentSizeObservationToken: NSKeyValueObservation {
-        if let observationToken = objc_getAssociatedObject(self, &UIScrollView.contentSizeObservableKey) as? NSKeyValueObservation {
+        if let observationToken = objc_getAssociatedObject(self, UIScrollView.contentSizeObservableKey.pointer) as? NSKeyValueObservation {
             return observationToken
         }
         let observationToken = observe(\.contentSize, options: .new) { [weak self] _, change in
@@ -54,17 +54,17 @@ extension UIScrollView {
             }
             self?.contentSizeObservable.on(.next(newValue))
         }
-        objc_setAssociatedObject(self, &UIScrollView.contentSizeObservationTokenKey, observationToken, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
+        objc_setAssociatedObject(self, UIScrollView.contentSizeObservationTokenKey.pointer, observationToken, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
 
         return observationToken
     }
     
     private var _contentSizeObservable: Observable<CGSize> {
-        if let observable = objc_getAssociatedObject(self, &UIScrollView.contentSizeObservableKey) as? Observable<CGSize> {
+        if let observable = objc_getAssociatedObject(self, UIScrollView.contentSizeObservableKey.pointer) as? Observable<CGSize> {
             return observable
         }
         let observable = Observable<CGSize>()
-        objc_setAssociatedObject(self, &UIScrollView.contentSizeObservableKey, observable, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
+        objc_setAssociatedObject(self, UIScrollView.contentSizeObservableKey.pointer, observable, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
         return observable
     }